Create another map page for scavenger hunt and add geolocation function
